% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data within tiles}
\usage{
aggregate(tile_data, geo_points, sum = NULL, count = F, crs)
}
\arguments{
\item{tile_data}{A spatial dataset (i.e. sf or SpatialPolygonsDataFrame) of polygon tiles}

\item{geo_points}{A spatial dataset (i.e. sf or SpatialPointsDataFrame) of points to aggregate}

\item{sum}{Variable from geo_points for calculating the sum per tile}

\item{count}{Count the number of points per tile. Input TRUE to count points. Defaults to FALSE}

\item{crs}{Coordinate Reference System}
}
\value{
An sf dataset containing the tiles with the aggregated data.
}
\description{
Function for aggregating data within map tiles.
}
\examples{
# Load the sf dataset of pumps and cholera deaths in Soho
data(soho_pumps, cholera_deaths)

# Generate tiles
tile_data <- propor::tiles(data = soho_pumps, shape = "hexagons")

# Aggregate data within tiles
propor::aggregate(tile_data = tile_data, geo_points = cholera_deaths, sum = "cholera.deaths", crs = 27700)
}
