% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propor.R
\name{propor}
\alias{propor}
\title{Data for Wurman-Style Maps}
\usage{
propor(data, w, max_val = NULL, max_var = NULL, crs = NULL)
}
\arguments{
\item{data}{A spatial dataset consisting of a grid of polygons}

\item{w}{Variable for weighting the data}

\item{max_val}{Value used to establish the weighting upper limit for sizing the cartogram polygons. Input either max_val or max_var}

\item{max_var}{Variable used to establish the weighting upper limit for sizing the cartogram polygons. max_var should not be the same as the weighting (w) parameter. Input either max_var or max_val}

\item{crs}{Coordinate Reference System}
}
\value{
An sf dataset containing the cartogram polygons and the original polygons.
}
\description{
Function for creating data for Wurman-style maps, a form of cartogram in which polygons (e.g. circles, hexagons, squares) arranged in a grid are proportionally resized according to a specified variable and visualised along with the original polygons for reference.
}
\examples{
# Load the sf dataset of housing stats
data(housing_london)

# Generate the data for the Wurman-style map
propor(housing_london, w = "new_dwellings", max_var = "lp2021_target")
}
