% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiregular.R
\name{semiregular}
\alias{semiregular}
\title{Build Semi-Regular Tessellating Tile Data}
\usage{
semiregular(data, shape = "circles_hexagons", crs = NULL)
}
\arguments{
\item{data}{A spatial dataset (i.e. sf or sp) for defining the bounding box}

\item{shape}{Shapes of the tiles to make. The only option currently is "circles_hexagons"}

\item{crs}{Coordinate Reference System}
}
\value{
An sf dataset containing the semi-regular tessellating tiles.
}
\description{
Function for creating semi-regular tessellating tiles including: a combination of circles and hexagons. Since the circles and hexagons tessellate when combined, they do not contain gaps unlike configurations that consist of only circles.
}
\examples{
# Load the sf dataset of Soho pumps
data(soho_pumps)

# Generate the tiles
propor::semiregular(data = soho_pumps)
}
