% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soho_pumps_data.R
\docType{data}
\name{soho_pumps}
\alias{soho_pumps}
\title{soho_pumps}
\format{
A data frame with 8 rows and 1 variables:
\describe{
\item{\code{soho.pump}}{character Soho pump}
\item{\code{geometry}}{sfc_POINT geometry}
}
}
\source{
\href{https://blog.rtwilson.com/john-snows-famous-cholera-analysis-data-in-modern-gis-formats/}{Robin's Blog}
}
\usage{
soho_pumps
}
\description{
An sf dataset containing the location of pumps in Soho from John Snow's analysis adapted from Robin Wilson's digitised dataset.
}
\details{
Data on Water Pumps in Soho
}
\keyword{datasets}
