% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tiles}
\alias{tiles}
\title{Build Regular Tessellating Tile Data}
\usage{
tiles(data, shape, crs = NULL)
}
\arguments{
\item{data}{A spatial dataset (i.e. sf or sp) for defining the bounding box}

\item{shape}{Shape of the tiles to make. Select either "triangles", "hexagons", "squares" or "diamonds"}

\item{crs}{Coordinate Reference System}
}
\value{
An sf dataset containing the tiles.
}
\description{
Function for creating simple meshes of regular tessellating map tiles - including triangles, hexagons, squares or diamonds - based on sf's st_make_grid function.
}
\examples{
# Load the sf dataset of Soho pumps
data(soho_pumps)

# Generate hexagonal tiles
propor::tiles(data = soho_pumps, shape = "hexagons")

# Generate triangular tiles
propor::tiles(data = soho_pumps, shape = "triangles")
}
