% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propor.R
\name{propor}
\alias{propor}
\title{Data for Wurman-Style Maps}
\usage{
propor(data, w, max_val = NULL, max_var = NULL, crs = NULL)
}
\arguments{
\item{data}{A spatial dataset (i.e. sf or SpatialPolygonsDataFrame) consisting of polygon tiles with data values}

\item{w}{The weighting variable}

\item{max_val}{Value used to establish the upper limit for the new polygons. Input either max_val or max_var}

\item{max_var}{Variable used to establish the upper limit for the new polygons. max_var should not be the same as the weighting variable. Input either max_var or max_val}

\item{crs}{Coordinate Reference System}
}
\value{
An sf dataset containing the new polygons and the original polygons.
}
\description{
Function for creating data for Wurman-style maps, a mapping technique in which tiles (e.g. circles, hexagons, squares etc) are proportionally resized according to a specified variable and visualised with the original tiles for comparison
}
\examples{
# Load the sf dataset of cholera deaths in Soho and pumps
data(soho_pumps, cholera_deaths)
# First, create hexagonal-shaped tiles
tile_data <- propor::tiles(data = soho_pumps, shape = "hexagons")
# Next, aggregate the data within the tiles
aggregated_data <- propor::aggregate(tile_data = tile_data, geo_points = cholera_deaths, sum = "cholera.deaths", crs = 27700)
# Finally, Generate the proportionally-resized tiles
propor::propor(aggregated_data, w = "cholera.deaths", max_val = 37)
}
